bot/pyproject.toml
[tool.poetry]
name = "rzd-bot"
version = "0.1.0"
description = "RZD Q&A Bot"
license = "Apache-2.0"
authors = ["AI RnD Lab Authors"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.13.1"
pyyaml = "^6.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
